# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: CCAT Practice App — Gateway API
  version: "1.0.0-phase0"
  description: >
    Normative /v1 API for the Concept Mastery CCAT Practice App. The Gateway is the ONLY
    application-data boundary exposed to the student app (Blueprint v10.0 §1.1, §33).
    Direct student Supabase/PostgREST access is prohibited (§41). All sensitive mutations
    use idempotency keys; all mutable Admin resources use expected-version/ETag concurrency
    (§22.4, §32.1). This document is Phase 0 and covers the endpoint surface of §32; request
    bodies for Admin write operations are represented by shared envelopes and will be
    expanded per-domain in Phase 1.
  contact:
    name: Concept Mastery
    email: admin@conceptmastery.ca
servers:
  - url: https://api.conceptmastery.ca/v1
    description: Production
  - url: https://staging-api.conceptmastery.ca/v1
    description: Staging (no real child PII, §38.1)
...
security:
  - studentSession: []
paths:
...
  /registration/contact/start:
    post:
      tags: [registration]
      operationId: registrationContactStart
      security: []
      summary: Begin guardian contact verification (§4.1, §32.2)
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel, destination]
              properties:
                channel: { $ref: "#/components/schemas/ContactChannel" }
                destination: { type: string, description: Email or E.164 phone }
      responses:
        "202": { description: OTP dispatched, $ref: "#/components/responses/ChallengeStarted" }
        "429": { $ref: "#/components/responses/RateLimited" }
        default: { $ref: "#/components/responses/Error" }
...
  /registration/consent:
    post:
      tags: [registration]
      operationId: registrationConsent
      security: []
      summary: Record immutable versioned guardian consent (§4.5)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [registration_grant, policy_version, consent_hash]
              properties:
                registration_grant: { type: string }
                policy_version: { type: string }
                consent_hash: { type: string }
      responses:
        "201": { description: Consent recorded }
        default: { $ref: "#/components/responses/Error" }
  /registration/student:
    post:
      tags: [registration]
      operationId: registrationStudent
      security: []
      summary: Create the student account after consent (§4.1, §4.3)
      parameters: [ { $ref: "#/components/parameters/IdempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudentRegistration" }
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudentProfile" }
        "422":
          description: Grade/age validation failed (§4.3)
          $ref: "#/components/responses/Error"
        default: { $ref: "#/components/responses/Error" }

  # -------------------------------------------------------------------- auth
...
components:
  securitySchemes:
    studentSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Student application session token, bound to the sole enrolled device (§5.4).
    adminSession:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin identity token (email/password + mandatory MFA, §22.1). Permissions loaded server-side.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, maxLength: 200 }
      description: Required on sensitive mutations (§13, §32.1). Retries with the same key return the original result.
...
  responses:
    Error:
      description: Structured error envelope (§32.1)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Rate limited (§36.4). Fails closed.
      headers:
        Retry-After: { schema: { type: integer } }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ChallengeStarted:
      description: OTP challenge created
      content:
        application/json:
          schema:
            type: object
            properties:
              challenge_id: { type: string, format: uuid }
              expires_at: { type: string, format: date-time }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, description: Machine code, e.g. ACTIVE_SESSION_EXISTS }
            message: { type: string }
            request_id: { type: string }
            details: { type: object, additionalProperties: true }
...
    ContactChannel: { type: string, enum: [email, sms] }
...
    StudentRegistration:
      type: object
      required: [registration_grant, display_name, username, grade_id, birth_month, birth_year, pin]
      properties:
        registration_grant: { type: string }
        display_name: { type: string }
        username: { type: string }
        grade_id: { type: string, format: uuid }
        birth_month: { type: integer, minimum: 1, maximum: 12 }
        birth_year: { type: integer }
        pin: { type: string, pattern: "^[0-9]{4}$" }
    StudentProfile:
      type: object
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        username: { type: string }
        grade_id: { type: string, format: uuid }
        age_years: { type: integer, description: Computed by Gateway from birth month/year (§4.2) }
        status: { $ref: "#/components/schemas/StudentStatus" }
        active_avatar_stage_id: { type: [string, "null"], format: uuid }
        active_theme_id: { type: [string, "null"], format: uuid }
    StudentStatus: { type: string, enum: [active, suspended, banned, pending_deletion, purged] }
...